+++ CHANGELOG.md
@@ -1,13 +1,15 @@
 == HEAD
 
+* Only add orphan elements to a `documentFragment` when necessary.
+
 == 1.0.2 (May 7, 20313)
 
-* Fix `matches()` on orphan elements in IE 9
+* Fix `matches()` on orphan elements in IE 9.
 
 == 1.0.1 (May 7, 2013)
 
-* Simplify AMD define
+* Simplify AMD define.
 
 == 1.0.0 (May 5, 2013)
 
-* Initial release
+* Initial release.